+++ cli/src/commands/dbs/delete.ts
@@ -27,14 +27,15 @@ export default class DatabasesDelete extends BaseCommand {
 
     const { confirm } = await this.prompt(
       {
-        type: 'confirm',
+        type: 'text',
         name: 'confirm',
-        message: `Are you sure you want to delete database ${database} in the ${workspace} workspace?`,
+        message: `Are you sure you want to delete database ${database} in the ${workspace} workspace? Please type ${database} to confirm`,
         initial: false
       },
-      flags.force
+      flags.force ? database : undefined
     );
     if (!confirm) return this.exit(1);
+    if (confirm !== database) return this.error('The database name did not match');
 
     await xata.databases.deleteDatabase(workspace, database);
